+++ lib/screens/polygraph/editors/horizontal_line_editor.dart
@@ -249,14 +249,14 @@ class _EditorLineHorizontalState extends ConsumerState<HorizontalLineEditor> {
                 child: TimeAggregationEditor()),
           ],
         ),
-        ElevatedButton(
-          onPressed: () {
-            setState(() {
-              /// blah
-            });
-          },
-          child: const Text('OK'),
-        ),
+        // ElevatedButton(
+        //   onPressed: () {
+        //     setState(() {
+        //       /// blah
+        //     });
+        //   },
+        //   child: const Text('OK'),
+        // ),
       ],
     );
   }