+++ lib/screens/polygraph/editors/transformed_variable_editor.dart
@@ -1,16 +1,12 @@
 library screens.polygraph.editors.transformed_variable_editor;
 
-import 'dart:ui';
 
 import 'package:flutter/material.dart' hide Interval, Transform;
-import 'package:flutter/scheduler.dart';
 import 'package:flutter_quiver/main.dart';
+import 'package:flutter_quiver/models/polygraph/polygraph_variable.dart';
+import 'package:flutter_quiver/models/polygraph/polygraph_window.dart';
 import 'package:flutter_quiver/models/polygraph/variables/transformed_variable.dart';
-import 'package:flutter_quiver/models/polygraph/transforms/transform.dart';
-import 'package:flutter_quiver/screens/polygraph/editors/editor_time_aggregation.dart';
-import 'package:flutter_quiver/screens/polygraph/editors/editor_time_filter.dart';
 import 'package:flutter_quiver/screens/polygraph/polygraph_window_ui.dart';
-import 'package:flutter_quiver/screens/polygraph/utils/autocomplete_field.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final providerOfTransformedVariable =
@@ -18,7 +14,11 @@ final providerOfTransformedVariable =
         (ref) => TransformedVariableNotifier(ref));
 
 class TransformedVariableEditor extends ConsumerStatefulWidget {
-  const TransformedVariableEditor({Key? key}) : super(key: key);
+  const TransformedVariableEditor(
+      {Key? key}) : super(key: key);
+
+  // final PolygraphWindow window;
+  // final bool pressedOk;
 
   @override
   ConsumerState<TransformedVariableEditor> createState() =>
@@ -84,8 +84,15 @@ class _TransformedVariableEditorState
     var state = ref.watch(providerOfTransformedVariable);
     controllerExpression.text = state.expression;
     controllerLabel.text = state.label;
+    // pressedOk = ref.watch(providerOfClickedOk).pushed;
+    // if (pressedOk) {
+    //   print('Validating........................................');
+    //   // validateForm(state);  // can't do this here, needs to run in an onPressed() method!
+    // }
     _errorMessage = state.error;
-    // print('_errorMessage: $_errorMessage');
+    print('_errorMessage: $_errorMessage');
+
+    /// TODO: Need to do validation on the change of tab too
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.end,
@@ -173,7 +180,7 @@ class _TransformedVariableEditorState
                             decoration: InputDecoration(
                               isDense: true,
                               contentPadding: const EdgeInsets.all(8),
-                              enabledBorder: (pressedOk &&
+                              enabledBorder: (state.error != '' &&
                                       controllerLabel.text == '')
                                   ? const OutlineInputBorder(
                                       borderSide: BorderSide(color: Colors.red))
@@ -220,13 +227,15 @@ class _TransformedVariableEditorState
                             decoration: InputDecoration(
                               isDense: true,
                               contentPadding: const EdgeInsets.all(8),
-                              enabledBorder: state.error != ''
+                              enabledBorder: (state.error != '' && controllerExpression.text == '')
                                   ? const OutlineInputBorder(
                                       borderSide: BorderSide(color: Colors.red))
                                   : InputBorder.none,
                             ),
                             onEditingComplete: () {
                               var window = ref.read(providerOfPolygraphWindow);
+                              print('in transformed_variable_editor, build, onEditingComplete:');
+                              print('${window.yVariables.map((e) => e.label)}');
                               setState(() {
                                 ref
                                     .read(
@@ -249,14 +258,14 @@ class _TransformedVariableEditorState
                     Row(
                       children: [
                         Container(
-                          width: 80,
+                          width: 100,
                           alignment: Alignment.centerRight,
                           padding: const EdgeInsets.only(right: 8),
                           child: const Text(''),
                         ),
                         Text(
-                          _errorMessage != '' ? 'Error: $_errorMessage' : '',
-                          style: const TextStyle(color: Colors.red),
+                          state.error != '' ? 'Error: ${state.error}' : '',
+                          style: const TextStyle(color: Colors.red, fontSize: 10),
                         ),
                       ],
                     ),
@@ -286,4 +295,23 @@ class _TransformedVariableEditorState
       ],
     );
   }
+
+  // void validateForm(TransformedVariable variable) {
+  //       var window = ref.read(providerOfPolygraphWindow);
+  //       setState(() {
+  //         pressedOk = true;
+  //         if (controllerLabel.text == '') {
+  //           variable.error =
+  //               'Label can\'t be empty';
+  //           return;
+  //         }
+  //         if (controllerExpression.text == '') {
+  //           variable.error = 'Expression can\'t be empty';
+  //           return;
+  //         }
+  //         variable.eval(window.cache);
+  //         var yVariables = [...window.yVariables, variable];
+  //         ref.read(providerOfPolygraphWindow.notifier).yVariables = yVariables;
+  //       });
+  // }
 }